+++ src/services/redux/slices/authSlice/authSelectors.js
@@ -1,3 +1,5 @@
 export const authStatusSelector = (state) => state.auth.status;
 
 export const authErrorSelector = (state) => state.auth.error;
+
+export const authUserSelector = (state) => state.auth.user;
